Zone player - multiroom audio player
This is the first checkin of my zone player.
The zone player is developed in c# is a poor mans version of a mulitroom
audio system.
The player has different players. Each of the players can be linked to a
different audio channel (e.g. USB devices). One can connect different
speakers to each audio channel and play or stream content in parallel to
the different audio channels